# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "终末地解谜 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/puzzles/balloon/recognize": {
      "post": {
        "summary": "Recognize Balloon",
        "operationId": "recognize_balloon_api_v1_puzzles_balloon_recognize_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "type": "object",
                "required": [
                  "image"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalloonRecognitionTaskView"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "408": {
            "description": "Request Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/puzzles/balloon/catalog/{code}": {
      "get": {
        "summary": "Balloon Catalog",
        "operationId": "balloon_catalog_api_v1_puzzles_balloon_catalog__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogEntry"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/puzzles/{puzzle_id}/recognize": {
      "post": {
        "summary": "Recognize",
        "operationId": "recognize_api_v1_puzzles__puzzle_id__recognize_post",
        "parameters": [
          {
            "name": "puzzle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Puzzle Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskView"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "image"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/puzzles/balloon/solve": {
      "post": {
        "summary": "Solve Balloon Route",
        "operationId": "solve_balloon_route_api_v1_puzzles_balloon_solve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BalloonPuzzle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalloonSolveTaskView"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/puzzles/{puzzle_id}/solve": {
      "post": {
        "summary": "Solve",
        "operationId": "solve_api_v1_puzzles__puzzle_id__solve_post",
        "parameters": [
          {
            "name": "puzzle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Puzzle Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskView"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "summary": "Get Task",
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskView"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      },
      "delete": {
        "summary": "Cancel Task",
        "operationId": "cancel_task_api_v1_tasks__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskView"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BalloonPuzzle": {
        "properties": {
          "rule_version": {
            "type": "string",
            "const": "center-torque-v1",
            "title": "Rule Version"
          },
          "rows": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 2.0,
            "title": "Rows"
          },
          "columns": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 2.0,
            "title": "Columns"
          },
          "usable_cells": {
            "items": {
              "$ref": "#/components/schemas/Cell"
            },
            "type": "array",
            "maxItems": 36,
            "minItems": 1,
            "title": "Usable Cells"
          },
          "inventory": {
            "items": {
              "$ref": "#/components/schemas/Stock"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 1,
            "title": "Inventory"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rule_version",
          "rows",
          "columns",
          "usable_cells",
          "inventory"
        ],
        "title": "BalloonPuzzle"
      },
      "BalloonRecognitionResult": {
        "properties": {
          "outcome": {
            "type": "string",
            "enum": [
              "draft",
              "no_board",
              "timeout",
              "invalid_image",
              "failed"
            ],
            "title": "Outcome"
          },
          "rows": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows"
          },
          "columns": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columns"
          },
          "cells": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "string",
                      "enum": [
                        "usable",
                        "blocked"
                      ]
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cells"
          },
          "inventory": {
            "items": {
              "$ref": "#/components/schemas/InventoryDraft"
            },
            "type": "array",
            "title": "Inventory"
          },
          "target_total_lift": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Total Lift"
          },
          "question_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Code"
          },
          "question_code_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Code Confidence"
          },
          "catalog": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CatalogMatch"
              },
              {
                "type": "null"
              }
            ]
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "outcome"
        ],
        "title": "BalloonRecognitionResult"
      },
      "BalloonRecognitionTaskView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "cancellation_requested": {
            "type": "boolean",
            "title": "Cancellation Requested",
            "default": false
          },
          "computation_stopped": {
            "type": "boolean",
            "title": "Computation Stopped",
            "default": false
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BalloonRecognitionResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at"
        ],
        "title": "BalloonRecognitionTaskView"
      },
...
      "BalloonSolveResult": {
        "properties": {
          "outcome": {
            "type": "string",
            "enum": [
              "solved",
              "unsatisfiable",
              "timeout"
            ],
            "title": "Outcome"
          },
          "rule_version": {
            "type": "string",
            "const": "center-torque-v1",
            "title": "Rule Version",
            "default": "center-torque-v1"
          },
          "solution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BalloonSolution"
              },
              {
                "type": "null"
              }
            ]
          },
          "limit_reason": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "time",
                  "work"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Reason"
          }
        },
        "type": "object",
        "required": [
          "outcome"
        ],
        "title": "BalloonSolveResult"
      },
      "BalloonSolveTaskView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "cancellation_requested": {
            "type": "boolean",
            "title": "Cancellation Requested",
            "default": false
          },
          "computation_stopped": {
            "type": "boolean",
            "title": "Computation Stopped",
            "default": false
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BalloonSolveResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at"
        ],
        "title": "BalloonSolveTaskView"
      },
      "CatalogCandidate": {
        "properties": {
          "fingerprint": {
            "type": "string",
            "title": "Fingerprint"
          },
          "puzzle": {
            "$ref": "#/components/schemas/BalloonPuzzle"
          },
          "target_total_lift": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Total Lift"
          },
          "observations": {
            "type": "integer",
            "title": "Observations",
            "default": 1
          },
          "first_seen": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen"
          }
        },
        "type": "object",
        "required": [
          "fingerprint",
          "puzzle",
          "first_seen",
          "last_seen"
        ],
        "title": "CatalogCandidate"
      },
      "CatalogEntry": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "status": {
            "type": "string",
            "enum": [
              "provisional",
              "verified",
              "disputed"
            ],
            "title": "Status"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CatalogCandidate"
            },
            "type": "array",
            "title": "Candidates"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "code",
          "status",
          "updated_at"
        ],
        "title": "CatalogEntry"
      },
...
      "Cell": {
        "properties": {
          "row": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Row"
          },
          "column": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Column"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "row",
          "column"
        ],
        "title": "Cell"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "InventoryDraft": {
        "properties": {
          "lift": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lift"
          },
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          }
        },
        "type": "object",
        "title": "InventoryDraft"
      },
...
      "Stock": {
        "properties": {
          "lift": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Lift"
          },
          "count": {
            "type": "integer",
            "maximum": 18.0,
            "minimum": 1.0,
            "title": "Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lift",
          "count"
        ],
        "title": "Stock"
      },
      "TaskView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "cancellation_requested": {
            "type": "boolean",
            "title": "Cancellation Requested",
            "default": false
          },
          "computation_stopped": {
            "type": "boolean",
            "title": "Computation Stopped",
            "default": false
          },
          "result": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at"
        ],
        "title": "TaskView"
      }
    }
  }
}